PA3: mostly working implementation, breaks backwards compatibility
"mostly works". still needs pretty-printing for compat with the rest
of the toolchain and error handling. also some small pieces still
rough (classParent, classFilename). I believe this breaks backwards
compatibility with the PA2 test suite because it prints a value for EOF
for debugging purposes. should probably rewrite to something like
"143pp" instead of show instances so i'm not tied to this interchange
format.